+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## In progress
 
+## 1.7.1
+
 - Vue progress/vow sheet ([#147](https://github.com/ben/foundry-ironsworn/pull/147))
 - Vue bondset sheet ([#148](https://github.com/ben/foundry-ironsworn/pull/148))
 - Vue helper for finding a sheet's actor ([#149](https://github.com/ben/foundry-ironsworn/pull/149))